pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-ripper"
version = "1.0.0"
description = "A production framework for modular deconstruction, analysis, and recomposition of knowledge in Transformer-based language models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
  { name = "qrv0", email = "qorvuscompany@gmail.com" }
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
  "torch>=2.0.0",
  "transformers>=4.30.0",
  "safetensors>=0.3.0",
  "h5py>=3.8.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "umap-learn>=0.5.6",
  "numba>=0.60.0; python_version >= '3.11'",
  "llvmlite>=0.43.0; python_version >= '3.11'",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "tqdm>=4.65.0",
  "datasets>=2.12.0",
  "accelerate>=0.20.0",
  "tokenizers>=0.13.0",
  "scipy>=1.10.0",
  "pandas>=2.0.0",
  "wandb>=0.15.0",
  "python-dotenv>=1.0.1",
  "nltk>=3.8.0",
  "spacy>=3.6.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.7.0", "seaborn>=0.12.0", "umap-learn>=0.5.6"]
spacy = ["spacy>=3.6.0"]
wandb = ["wandb>=0.15.0"]

[project.scripts]
llm-ripper = "llm_ripper.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/llm-ripper"
Repository = "https://github.com/your-org/llm-ripper"

[tool.ruff]
line-length = 100
target-version = "py38"
extend-exclude = ["venv", "src/llm_ripper.egg-info"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["F401", "F841", "F541", "E501"]
"src/llm_ripper/studio/*" = ["F401", "F841"]
"src/llm_ripper/features/*" = ["F401", "F841"]
"src/llm_ripper/interop/*" = ["F401", "F841"]
"src/llm_ripper/safety/*" = ["F401", "F841"]
"src/llm_ripper/uq/*" = ["F401", "F841"]
"src/llm_ripper/utils/visualization.py" = ["F401", "F841"]
"src/llm_ripper/utils/hdf5_tools.py" = ["F401"]
"src/llm_ripper/utils/metrics.py" = ["F401"]
"src/llm_ripper/utils/model_loader.py" = ["F401"]
"tests/*" = ["F401", "F841", "E702"]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
no_implicit_optional = false
exclude = "venv|src/llm_ripper/utils/visualization.py"
